+++ src/interfaces/IMasterChef.sol
@@ -3,27 +3,24 @@
 pragma solidity 0.6.12;
 
 interface IMasterChef {
-  function sushi() external view returns (address);
+    function sushi() external view returns (address);
 
-  function poolInfo(uint pid)
-    external
-    view
-    returns (
-      address lpToken,
-      uint allocPoint,
-      uint lastRewardBlock,
-      uint accSushiPerShare
-    );
+    function poolInfo(uint256 pid)
+        external
+        view
+        returns (
+            address lpToken,
+            uint256 allocPoint,
+            uint256 lastRewardBlock,
+            uint256 accSushiPerShare
+        );
 
-  function userInfo(uint pid, address user)
-    external
-    view
-    returns (
-      uint amount,
-      uint rewardDebt
-    );
+    function userInfo(uint256 pid, address user)
+        external
+        view
+        returns (uint256 amount, uint256 rewardDebt);
 
-  function deposit(uint pid, uint amount) external;
+    function deposit(uint256 pid, uint256 amount) external;
 
-  function withdraw(uint pid, uint amount) external;
+    function withdraw(uint256 pid, uint256 amount) external;
 }
