+++ src/interfaces/IWETH.sol
@@ -3,9 +3,13 @@
 pragma solidity 0.6.12;
 
 interface IWETH {
-  function balanceOf(address user) external returns (uint);
-  function approve(address to, uint value) external returns (bool);
-  function transfer(address to, uint value) external returns (bool);
-  function deposit() external payable;
-  function withdraw(uint) external;
+    function balanceOf(address user) external returns (uint256);
+
+    function approve(address to, uint256 value) external returns (bool);
+
+    function transfer(address to, uint256 value) external returns (bool);
+
+    function deposit() external payable;
+
+    function withdraw(uint256) external;
 }
